Tests moved inside Pacakges (idamApi,rdProfessionalApi. Removed console.log
